+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.2.1 - 2020-04-30
+
+### Added
+
+- A notice will now be shown in the admin area if a refund flow is used that requires approval from the merchant.
+
+### Changed
+
+- Changed version checker behavior from flag storage to cache storage.
+
+### Fixed
+
+- Fixes issue where the title of the payment method was not shown in the order grid if the order was paid with Ingenico.
+- Fixes issue where a cancellation on the RPP would not cancel the order in Magento.
+
 ## 2.2.0.1 - 2020-04-09
 
 ### Fixed
@@ -151,6 +166,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 1.5.0 - 2018-12-14
 
 ### Added
+
 - config option to specify hosted checkout variant
 - ability for a full redirect with payment product selection on hosted checkout
 - ability for a custom prefix as system identifier for merchantReference used to identify orders
@@ -160,12 +176,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - support for Magento 2.3
 
 ### Changed
+
 - webhook endpoints now always return a 200 code, if the event payload could be unwrapped
 - ensure every order has a transaction with the reference to Ingenico's payId if a response was received
 - Invoice for CAPTURE_REQUESTED status on CC is now in Pending state rather then Paid
 - streamline Ingenico status processing notifications in the order history
 
 ### Fixed
+
 - order status inconsistencies with credit card
 - order payment accept action not available on fraud status
 - order not progressing from payment review order state to processing
@@ -173,11 +191,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 1.4.5 - 2018-11-02
 
 ### Fixed
+
 - Orders in suspected fraud state could not be accepted/denied
 
 ## 1.4.4 - 2018-10-05
 
 ### Fixed
+
 - Automatic Order Update does not retrieve payment ID correctly
 - Payment methods not reloading when removing coupon in checkout (EE)
 - Fatal error during order placement in Mage 2 EE
@@ -190,34 +210,42 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 1.4.3 - 2018-09-12
 
 ### Fixed
+
 - Inline card payments that result in a redirect will now be handled correctly
 
 ### Changed
+
 - Extension copyright now lies with Ingenico eCommerce Solutions Bvba
 
 ## 1.4.2 - 2018-08-27
 
 ### Added
+
 - automatically send invoice email after it is marked as paid
 
 ### Changed
+
 - extension is now licensed under MIT (see LICENSE.txt)
 
 ### Fixed
+
 - fixed issue with refilling carts where product was not loaded properly
 - corrected formatting for customer date of birth
 
 ## 1.4.1 - 2018-08-06
 
 ### Fixed
+
 - restored compatibility with Magento 2.1
 
 ## 1.4.0 - 2018-08-02
 
 ### Added
+
 - automatic configuration validation against API when saving changed account settings
 
 ### Changed
+
 - automatically update payment status from API after invoice cancellation
 - status handling to be more flexible by implementing simple status handlers
 - Credit Card orders on Global Collect can be shipped with status 800,900,975 
@@ -226,9 +254,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deprecated
 
 ### Removed
+
 - WX file import admin and cron configuration
 
 ### Fixed
+
 - invoice cancellation not working properly
 - customer gender not transmitted as string
 - incompatibility with CheckoutAgreements core extension
@@ -237,47 +267,68 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 
 ## 1.3.1 - 2018-06-05
+
 ### Changed
+
 - adjusted order item transmission to be compatible with more tax calculation settings
 
 ## 1.3.0 - 2018-05-30
+
 ### Added
+
 - added compatibility with OneStepCheckout](http://onestepcheckout.com) extension
+
 ### Changed
+
 - changed handling of returning customers from HostedCheckout page to no longer rely on the checkout session
 - replaced usages of deprecated Cart interface
+
 ### Removed
+
 - removed transmission and reliance on Magento order entity id for identifying an order, using the increment id in all places instead
+
 ###Fixed
+
 - fixed an issue which occured if the module code was already present during Magento installation
 
 ## 1.2.0 - 2018-04-20
+
 ### Added
+
 - Inline payment workflow to allow direct payment creation without redirects for supporting payment products
 - WX file polling automatically retrieves the daily transaction report file in xml format and parses the updates into Magentos orders
 
 ## 1.1.0 - 2018-03-21
+
 ### Changed
+
 - updated order item transmission to allow better display on HostedCheckout page for shipping and discount amounts
 - integrate Ingenicos Javascript SDK to fetch the available payment products through the client
 - improved error handling during checkout to enable more graceful checkout expirience
 - make API endpoints required in the configuration
 
 ### Fixed
+
 - fixed wrong redirection for customers returning from HostedCheckout
 
 ## 1.0.1 - 2018-02-22
+
 ### Added
+
 - handling for AUTHORIZATION_REQUESTED API status
 - PHP 7.1 compatibility by removal of preserved keywords
 
 ### Changed
+
 - fallback to fetch hostedCheckout status, if no paymentId is present at the order yet
 - handle possible API errors in transaction information to be readable
 
 ### Fixed
+
 - bug in automatic order cancellation to not fetch all necessary orders
 
 ## 1.0.0 - 2018-01-16
+
 ### Added
+
 - Initial release